benchmarks: fix narrowing warnings by clang-tidy

+++ test/benchmark_conveyor_belt.cpp
@@ -149,12 +149,14 @@ BM_ConveyorBelt(benchmark::State &state, bool weighted = true, bool multi_thread
 		controller_size += controller.get_locations().size();
 	}
 
-	state.counters["tree_size"] = benchmark::Counter(tree_size, benchmark::Counter::kAvgIterations);
+	state.counters["tree_size"] =
+	  benchmark::Counter(static_cast<double>(tree_size), benchmark::Counter::kAvgIterations);
 	state.counters["pruned_tree_size"] =
-	  benchmark::Counter(pruned_tree_size, benchmark::Counter::kAvgIterations);
+	  benchmark::Counter(static_cast<double>(pruned_tree_size), benchmark::Counter::kAvgIterations);
 	state.counters["controller_size"] =
-	  benchmark::Counter(controller_size, benchmark::Counter::kAvgIterations);
-	state.counters["plant_size"] = benchmark::Counter(plant_size, benchmark::Counter::kAvgIterations);
+	  benchmark::Counter(static_cast<double>(controller_size), benchmark::Counter::kAvgIterations);
+	state.counters["plant_size"] =
+	  benchmark::Counter(static_cast<double>(plant_size), benchmark::Counter::kAvgIterations);
 }
 
 BENCHMARK_CAPTURE(BM_ConveyorBelt, single_heuristic, false)
